+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/plugins/project/IntelliJPlatformModulePlugin.kt
@@ -189,6 +189,8 @@ abstract class IntelliJPlatformModulePlugin : Plugin<Project> {
                 Configurations.INTELLIJ_PLATFORM_TEST_CLASSPATH,
                 Configurations.INTELLIJ_PLATFORM_TEST_RUNTIME_CLASSPATH,
                 Configurations.INTELLIJ_PLATFORM_RUNTIME_CLASSPATH,
+                Configurations.INTELLIJ_PLATFORM_SANDBOX_RUNTIME_CLASSPATH,
+                Configurations.INTELLIJ_PLATFORM_TEST_SANDBOX_RUNTIME_CLASSPATH,
                 Configurations.External.COMPILE_CLASSPATH,
                 Configurations.External.TEST_COMPILE_CLASSPATH,
                 Configurations.External.TEST_RUNTIME_CLASSPATH,